+++ src/screens/tela_lista_alunos/index.tsx
@@ -103,21 +103,25 @@ export default function TelaListaAlunos(){
      */
      const deletarCadastro = async (idAluno: string, posicaoArray: number) => {
         try{
-            setStatusCarregando("Deletando cadastro...");
+            let confirmacao = window.confirm("Deseja mesmo deletar o aluno e seus treinos?");
 
-            //Deleta o documento do aluno
-            await deleteDoc(doc(db, "alunos", idAluno));
+            if(confirmacao === true){
+                setStatusCarregando("Deletando cadastro...");
 
-            //Deleta os treinos do aluno
-            let colecaoTreinos = await getDocs(collection(db, "alunos", idAluno, "treinos"));
-            colecaoTreinos.forEach(documento => {
-                deleteDoc(documento.ref);
-            });
-
-            //Deleta do vetor de alunos para que não precise fazer a consulta novamente no banco de dados
-            let listaAlunos = alunos;
-            listaAlunos.splice(posicaoArray, 1);
-            setAlunos([...listaAlunos]);
+                //Deleta o documento do aluno
+                await deleteDoc(doc(db, "alunos", idAluno));
+    
+                //Deleta os treinos do aluno
+                let colecaoTreinos = await getDocs(collection(db, "alunos", idAluno, "treinos"));
+                colecaoTreinos.forEach(documento => {
+                    deleteDoc(documento.ref);
+                });
+    
+                //Deleta do vetor de alunos para que não precise fazer a consulta novamente no banco de dados
+                let listaAlunos = alunos;
+                listaAlunos.splice(posicaoArray, 1);
+                setAlunos([...listaAlunos]);
+            }
         }catch(erro){
             alert(erro);
         }finally{
